+++ paper.bib
@@ -2,7 +2,8 @@
   title={Air pollution health impacts: the knowns and unknowns for reliable global burden calculations},
   author={Hystad, Perry and Yusuf, Salim and Brauer, Michael},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/cvr/cvaa092}
 }
 
 @article{Hyndman:2008,
@@ -54,6 +55,7 @@
   volume={5},
   number={50},
   pages={2250},
-  year={2020}
+  year={2020},
+  doi={10.21105/joss.02250},
 }
 
