+++ paper.bib
@@ -56,6 +56,6 @@
   number={50},
   pages={2250},
   year={2020},
-  doi={10.21105/joss.02250},
+  doi={10.21105/joss.02250}
 }
 
