+++ paper.bib
@@ -1,9 +1,9 @@
 @misc{Hystad:2020,
-  title={Air pollution health impacts: the knowns and unknowns for reliable global burden calculations},
-  author={Hystad, Perry and Yusuf, Salim and Brauer, Michael},
-  year={2020},
-  publisher={Oxford University Press},
-  doi={10.1093/cvr/cvaa092}
+  title = {Air pollution health impacts: the knowns and unknowns for reliable global burden calculations},
+  author = {Hystad, Perry and Yusuf, Salim and Brauer, Michael},
+  year = {2020},
+  publisher = {Oxford University Press},
+  doi = {10.1093/cvr/cvaa092}
 }
 
 @article{Hyndman:2008,
@@ -15,6 +15,7 @@
     pages = {1--22},
     year = {2008},
     url = {https://www.jstatsoft.org/article/view/v027i03},
+    doi = {10.18637/jss.v027.i03}
   }
 
 @manual{Shiny:2021,
@@ -38,24 +39,24 @@
 }
 
 @article{Salmon:2017,
-  title={rtimicropem: an R package supporting the analysis of RTI MicroPEM output files},
-  author={Salmon, Ma{\"e}lle and Vakacherla, Sreekanth and Mil{\`a}, Carles and Marshall, Julian and Tonne, Cathryn},
-  journal={Journal of Open Source Software},
-  volume={2},
-  number={16},
-  pages={333},
-  year={2017},
-  doi={10.21105/joss.00333}
+  title = {rtimicropem: an R package supporting the analysis of RTI MicroPEM output files},
+  author = {Salmon, Ma{\"e}lle and Vakacherla, Sreekanth and Mil{\`a}, Carles and Marshall, Julian and Tonne, Cathryn},
+  journal = {Journal of Open Source Software},
+  volume = {2},
+  number = {16},
+  pages = {333},
+  year = {2017},
+  doi = {10.21105/joss.00333}
 }
 
 @article{Upadhya:2020,
-  title={mmaqshiny v1. 0: R-Shiny package to explore Air-Quality Mobile-Monitoring data},
-  author={Upadhya, Adithi R and Agrawal, Pratyush and Vakacherla, Sreekanth and Kushwaha, Meenakshi},
-  journal={Journal of Open Source Software},
-  volume={5},
-  number={50},
-  pages={2250},
-  year={2020},
-  doi={10.21105/joss.02250}
+  title = {mmaqshiny v1. 0: R-Shiny package to explore Air-Quality Mobile-Monitoring data},
+  author = {Upadhya, Adithi R and Agrawal, Pratyush and Vakacherla, Sreekanth and Kushwaha, Meenakshi},
+  journal = {Journal of Open Source Software},
+  volume = {5},
+  number = {50},
+  pages = {2250},
+  year = {2020},
+  doi = {10.21105/joss.02250}
 }
 
