### 2021-06-30 #3435 @whedon
[REVIEW]: pollucheck v1.0: A package to explore open-source air pollution data
**Submitting author:** @adithirugis (<a href="http://orcid.org/0000-0002-1764-1379">Adithi Upadhya</a>)
**Repository:** <a href="https://github.com/adithirgis/pollucheck" target ="_blank">https://github.com/adithirgis/pollucheck</a>
**Version:** 1.0.0
