+++ README.md
@@ -2,13 +2,10 @@
 
 PolluCheck helps exploring the open source Indian air quality data.
 
-⚡ Fun fact: Its a tidy code!
-
-The link for this app is this - <https://aruapps.shinyapps.io/OpenSourceAirQualityApp/>.
-
 -   PolluCheck (currently in the development phase) is specific to India, and allows users to handle open-source air-quality datasets available from OpenAQ (<https://openaq.org/#/countries/IN?_k=5ecycz>), CPCB (<https://app.cpcbccr.com/ccr/#/caaqm-dashboard-all/caaqm-landing>), and AirNow (<https://www.airnow.gov/international/us-embassies-and-consulates/#India>).
 -   Users can visualize data, analyse data, perform basic statistical operations, and generate a variety of publication-ready plots.
 -   We have also included the popular openair package in this application.
+-   We have hosted this app here - https://aruapps.shinyapps.io/OpenSourceAirQualityApp/ 
 
 # Have a look at the app!
 
