### 2021-07-08 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Review for joss
hello @adithirgis, how are you? I hope everything is fine.

I`m starting this issue to review your article, according to the guidelines suggestions as mentioned here https://github.com/openjournals/joss-reviews/issues/3435

I will be editing this issue and adding more items as my review progress.

### [Checklist The JOSS paper](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper)

- [x] A list of the authors of the software and their affiliations.
- [x] A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
- [x] A clear statement of need that illustrates the purpose of the software.
- [x] A description of how this software compares to other commonly-used packages in this research area.
- [x] Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it.
- [x] A list of key references including a link to the software archive.
- [x] Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

### [Checklist license](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license)
- [x] Acceptable: A plain-text LICENSE or COPYING file with the contents of an OSI approved license

### [Checklist Substantial scholarly effort](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort)

- [x] Age of software (is this a well-established software project) / length of commit history.
- [x] Number of commits.
- [x] Number of authors.
- [x] Lines of code (LOC): These statistics are usually reported by Whedon in the pre-review issue thread.
- [x] Whether the software has already been cited in academic papers.
- [x] Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain.

### [Statement of need](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need)

- [x] The authors should clearly state what problems the software is designed to solve and who the target audience is

### [Tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)

- [x] An automated test suite hooked up to an external service such as Travis-CI or similar

### [Autorship](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship)

- [x] Indicate what authors did
