+++ paper.md
@@ -22,55 +22,55 @@ affiliations:
 - index: '1'
   name: ILK Labs, Bengaluru, India
 - index: '2'
-  name: Center for Study of Science, Technology & Policy, Bengaluru, India
+  name: Center for Study of Science, Technology and Policy, Bengaluru, India
 ---
 
 # Summary
 
-Air pollution is one of the major environmental threats impacting human health, quality of living, climate and economy [@Hystad:2020]. Quantification of the pollution is needed to assess its impact on the above said aspects and device mitigation actions. Measurements are the most accurate way of quantification of air pollution. Many countries conduct regulatory measurements of various air pollutants (e.g. fine and respirable particulate matter, nitrogen dioxide, sulfur dioxide, surface ozone etc.) and make the data available publicly.
+Air pollution impacts human health, quality of living, climate and the economy [@Hystad:2020]. To assess its impact and device mitigation actions quantification of air pollution is vital. Measurements are the most accurate way of quantifying air pollution. Many countries conduct regulatory measurements of various air pollutants (e.g, fine and respirable particulate matter, nitrogen dioxide, sulfur dioxide and surface ozone) and make the data available publicly.
 
-Air pollution datasets typically span several seasons, years and the real-time data is recorded typically every hour or a higher frequency. With ever increasing quantum of data and number of data providers, there is a clear need for tools to handle, analyse and visualise large datasets. `Shiny` is a package in R which helps in building interactive applications [@Shiny:2021]. The current shiny app `pollucheck` aims at a simple workflow to generate a suite of statistical plots and summary statistics; users do not need any programming background to be able to analyse time-series data and generate a variety of plots.
+Air pollution data sets typically span several seasons or years and real-time data are recorded typically every hour or at a higher frequency. With the ever increasing quantum of data and number of data providers, there is a clear need for tools to handle, analyse, and visualise large data sets. `Shiny` is a package in R that helps build interactive applications [@Shiny:2021]. The current Shiny app `pollucheck` aims at a simple workflow to generate a suite of statistical plots and summary statistics. Users do not need any programming background to analyse time-series data and generate a variety of plots.
 
-`pollucheck` can handle real-time pollution and colocated meteorological (if available) data from three most popular open source air pollution databases. The data sources include [OpenAQ](openaq.org), [Airnow](airnow.gov) and [Indian Central Pollution Control Board (CPCB) dashboard](app.cpcbccr.com). While CPCB data is specific to Indian regulatory monitoring stations, OpenAQ hosts the global open source pollution databases and Airnow hosts the global PM~2.5~ (mass concentration of particulate matter with aerodynamic diameter less than or equal to 2.5 microns) data being collected under the United States Embassy and Consulates' air quality monitoring program.
+`pollucheck` can handle real-time pollution and co-located meteorological data (if available) from the three most popular open-source air pollution databases: [OpenAQ](openaq.org), [AirNow](airnow.gov) and [Indian Central Pollution Control Board (CPCB) dashboard](app.cpcbccr.com). While CPCB data are specific to Indian regulatory monitoring stations, OpenAQ hosts the global open source pollution databases and AirNow hosts the global PM~2.5~ (mass concentration of particulate matter with an aerodynamic diameter less than or equal to 2.5 microns) data, collected under the United States Embassy and Consulates' air quality monitoring programmes.
 
-Pollution data from the above sources is typically in different file formats and templates requiring customised codes/programs for analysis. Also a rigorous quality check of these data is preferred before visualization (plotting) and/or reporting. `pollucheck` offers a single stop solution for
[…]
