### 2021-07-14 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Hello @adithirgis

i have been trying to download from CCR but I can't. I only got a message saying "Processing". I think you package is very useful and I'm wondering if I can use with my own data.

![Screenshot from 2021-07-14 12-37-15](https://user-images.githubusercontent.com/27447280/125651316-65ec0dd5-955c-45f0-9f6e-81a1428d6afc.png)

Do you have a sample data?

Then, I downloaded your data and open it with LibreOffice, finding that it includes row.names. Did you use row.names = T when you generate the csv?
![Screenshot from 2021-07-14 12-43-05](https://user-images.githubusercontent.com/27447280/125651559-42916c20-36cf-407a-9896-d73f370577d6.png)
