### 2021-07-14 adithirgis/pollucheck#1 @adithirgis
Hello!

Thank you for reviewing the application.
1. Sorry the site where you download CCR, sometimes has issues, you can try
clicking the excel option again and see if it downloads. I am attaching a
few hourly files from CCR and from OpenAir and AirNow to explore.
2. I have not added *row.names = TRUE*. I could take a look into that.

Cheers!
Adithi

On Wed, Jul 14, 2021 at 9:15 PM Sergio Ibarra Espinosa <
***@***.***> wrote:

> Hello @adithirgis <https://github.com/adithirgis>
>
> i have been trying to download from CCR but I can't. I only got a message
> saying "Processing". I think you package is very useful and I'm wondering
> if I can use with my own data.
>
> [image: Screenshot from 2021-07-14 12-37-15]
> <https://user-images.githubusercontent.com/27447280/125651316-65ec0dd5-955c-45f0-9f6e-81a1428d6afc.png>
>
> Do you have a sample data?
>
> Then, I downloaded your data and open it with LibreOffice, finding that it
> includes row.names. Did you use row.names = T when you generate the csv?
>
