### 2021-07-14 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hello!

Thank you for reviewing the application.
1. Sorry the site where you download CCR, sometimes has issues, you can try clicking the excel option again and see if it downloads. I am adding a few hourly files [here](https://github.com/adithirgis/pollucheck/tree/main/inst/shiny/data_check) from CCR and from OpenAir and AirNow to explore.
2. I have not added `row.names = TRUE`. I could take a look into that.

Cheers!
Adithi
