### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hi,

Just wondering if we have have multiple reviews here and in the [issues of JOSS](https://github.com/openjournals/joss-reviews/issues/3435).

Thanks!
