### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Hi,

Just one review. Here is more like a personal review based on JOSS
guidelines. The official review is here
https://github.com/openjournals/joss-reviews/issues/3435.

Em qui., 15 de jul. de 2021 às 01:34, Adithi R. Upadhya <
***@***.***> escreveu:

> Hi,
>
> Just wondering if we have have multiple reviews here and in the issues of
> JOSS.
>
> Thanks!
>
