### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
### Observations of the manuscript:

General: This a great work and I'm interested in sharing with colleagues who work with observations.  The only issues us that the manuscript shows a description of the API rather than paper. In this sense it would be nice to add an example and its explanations. For instance, on Line 53 you mention that the ratio PM2.5/PM10 is a useful metric on air pollution field but you do not mention why. Then, my suggestion is to add a section analyzing a case of air pollution, using and explaining some of the tools that you have developed. Again, this is a suggestion only.

L8: I don't understand the word *device*. You mean promote?
L13: Is quantum appropriate?
