### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
I`m having lots of fun with this package. AIRNOW works perfectly, but I cant upload OpenAQ data. What are the restrictions, min 15-min?
