### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Add a [CrediT](https://www.elsevier.com/authors/policies-and-guidelines/credit-author-statement) statement mentioning what authors did
