### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Please, add tests.
