### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hello,

Thanks for the review, here are the tests - https://github.com/adithirgis/pollucheck/tree/main/inst/shiny/tests/shinytest.
