### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> Add a CrediT statement mentioning what authors did

We have added the roles of the authors in the [DESCRIPTION](https://github.com/adithirgis/pollucheck/blob/main/DESCRIPTION) file, do we need to add anymore?
