### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> I`m having lots of fun with this package. AIRNOW works perfectly, but I cant upload OpenAQ data. What are the restrictions, min 15-min?

So, the options/restrictions for CCR are 15, 30 and 60 min only, while for OpenAQ there are no options or restrictions, I am afraid if the CCR option is checked.
