### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
> > Add a CrediT statement mentioning what authors did
>
> We have added the roles of the authors in the [DESCRIPTION](https://github.com/adithirgis/pollucheck/blob/main/DESCRIPTION) file, do we need to add anymore?

> > I`m having lots of fun with this package. AIRNOW works perfectly, but I cant upload OpenAQ data. What are the restrictions, min 15-min?
>
> So, the options/restrictions for CCR are 15, 30 and 60 min only, while for OpenAQ there are no options or restrictions, I am afraid if the CCR option is checked.

[measurements_2.csv](https://github.com/adithirgis/pollucheck/files/6824584/measurements_2.csv)
I cant make it work with this file
