### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hello!

Were you able to work with the CCR data?

And was I able to respond to these following comments correctly? (Sorry I am a little confused)

> Please, add tests.

> Add a CrediT statement mentioning what authors did

Thanks!
