### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hello,

>measurements_2.csv
I cant make it work with this file

Thanks so much for pointing it out. I hope it works now :)
