### 2021-07-15 202c55f
small update - quantum and device
