### 2021-07-15 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> L8: I don't understand the word device. You mean promote?
L13: Is quantum appropriate?

I have changed device to **_facilitate_** and quantum to **_amount_**.
