### 2021-07-16 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> Add a CrediT statement mentioning what authors did

Where should I add this statement?

Adithi R. Upadhya created the package and contributed to the manuscript and she will maintain the package with Meenakshi Kushwaha who also supervised and contributed to the manuscript, Pratyush Agrawal contributed to designing, testing the application and data collection while Sreekanth Vakacherla designed, contributed to the manuscript and supervised the project.
