### 2021-07-16 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> Line 53 you mention that the ratio PM2.5/PM10 is a useful metric on air pollution field but you do not mention why.

We have added the reasons along with references in manuscript, thank you for pointing it out.

"the app computes the PM2.5/PM10 ratio, a useful metric in the air pollution field to identify sources of PM and to estimate PM2.5 when only PM10 is available [Chan:2008; Chu:2015; Spandana:2021]."
