+++ paper.bib
@@ -59,4 +59,37 @@
   year = {2020},
   doi = {10.21105/joss.02250}
 }
+@article{Spandana:2021,
+  title = {PM2. 5/PM10 ratio characteristics over urban sites of India},
+  author = {Spandana, B and Rao, S Srinivasa and Upadhya, Adithi R and Kulkarni, Padmavati and Sreekanth, V},
+  journal = {Advances in Space Research},
+  volume = {67},
+  number = {10},
+  pages = {3134--3146},
+  year = {2021},
+  publisher = {Elsevier},
+  doi = {10.1016/j.asr.2021.02.008}
+}
 
+@article{Chan:2008,
+  title = {Air pollution in mega cities in China},
+  author = {Chan, Chak K and Yao, Xiaohong},
+  journal = {Atmospheric environment},
+  volume = {42},
+  number = {1},
+  pages = {1--42},
+  year = {2008},
+  publisher = {Elsevier},
+  doi = {10.1016/j.atmosenv.2007.09.003}
+}
+
+@article{Chu:2015,
+  title = {Modeling the spatio-temporal heterogeneity in the PM10-PM2. 5 relationship},
+  author = {Chu, Hone-Jay and Huang, Bo and Lin, Chuan-Yao},
+  journal = {Atmospheric Environment},
+  volume = {102},
+  pages = {176--182},
+  year = {2015},
+  publisher = {Elsevier},
+  doi = {10.1016/j.atmosenv.2014.11.062}
+}
