### 2021-07-16 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
> my suggestion is to add a section analyzing a case of air pollution, using and explaining some of the tools that you have developed.

We are working on this. Thank you
