+++ README.md
@@ -4,9 +4,9 @@
 [![R-CMD-check](https://github.com/adithirgis/pollucheck/workflows/R-CMD-check/badge.svg)](https://github.com/adithirgis/pollucheck/actions)
  
 
-PolluCheck helps exploring the open source Indian air quality data.
+pollucheck helps exploring the open source Indian air quality data.
 
--   PolluCheck (currently in the development phase) is specific to India, and allows users to handle open-source air-quality datasets available from OpenAQ (<https://openaq.org/#/countries/IN?_k=5ecycz>), CPCB (<https://app.cpcbccr.com/ccr/#/caaqm-dashboard-all/caaqm-landing>), and AirNow (<https://www.airnow.gov/international/us-embassies-and-consulates/#India>).
+-   pollucheck is specific to India, and allows users to handle open-source air-quality datasets available from OpenAQ (<https://openaq.org/#/countries/IN?_k=5ecycz>), CPCB (<https://app.cpcbccr.com/ccr/#/caaqm-dashboard-all/caaqm-landing>), and AirNow (<https://www.airnow.gov/international/us-embassies-and-consulates/#India>).
 -   Users can visualize data, analyse data, perform basic statistical operations, and generate a variety of publication-ready plots.
 -   We have also included the popular openair package in this application.
 -   We have hosted this app here - https://aruapps.shinyapps.io/OpenSourceAirQualityApp/ 
