### 2021-07-16 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
It would be nice to identify the units on the data-base and add it on the plot, keeping the option Edit Y axis label
