### 2021-07-17 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
I am sorry but some files downloaded from CCR does not have units attached in the raw file. Any suggestions we should hard code it?

Thanks!
