### 2021-07-17 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
In that case it is hard. However, if it is possible, you could make default
reading the units from the database, and if the units is not available, the
you could just add "units" or something. I will check your code

Em sáb, 17 de jul de 2021 01:58, Adithi R. Upadhya ***@***.***>
escreveu:

> I am sorry but some files downloaded from CCR does not have units attached
> in the raw file. Any suggestions we should hard code it?
>
> Thanks!
>
