### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Hello @ibarraespinosa,

Should I change the Y axis label values everywhere instead of **Y parameter**  to **Y parameter (units)**?

And also have we been able to respond to your previous comments?

Thanks
Adithi
