### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
Hello Adithirgis,

I am not sure. It would be fine that your program detect automatically the
untis. If that is not possible you can keep it as it is. YEah, all your
comments are good so far.

Have a nice week.

Best, Sergio

Em seg., 19 de jul. de 2021 às 02:57, Adithi R. Upadhya <
***@***.***> escreveu:

> Hello @ibarraespinosa <https://github.com/ibarraespinosa>,
>
> Should I change the Y axis label values everywhere instead of *Y
> parameter* to *Y parameter (units)*?
>
> And also have we been able to respond to your previous comments?
>
> Thanks
> Adithi
>
