### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Thanks for the quick response.

I was trying to make it work, but it seems like it would be difficult to add for CCR files. I hope this is ok. Thanks so much for your review. We really appreciate it.

Where should we add the credit statement?

Also we are almost completing with the case study, I will update in the paper.md soon.
