### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
In the paper.md or in the ReadMe file?

We will wait for the review to finish in https://github.com/openjournals/joss-reviews/issues/3435

Thanks
