### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @ibarraespinosa
> In the paper.md or in the ReadMe file?
>
> We will wait for the review to finish in [openjournals/joss-reviews#3435](https://github.com/openjournals/joss-reviews/issues/3435)
>
> Thanks

For me it is fine, the paper and the software are already very good. Just try to fix the problem with the upload of openAQ.
For me the revision is over.

You have developed another great software to address air pollution problems. I hope you can publish many papers using pollucheck. I`m sure it will be used around the globe.

Best, Sergio
