### 2021-07-19 #3435 @ibarraespinosa
Hello  @ajstewartlang, I have already reviewed the paper by @adithirgis. It consists in a software to analyze air pollution measurements with different databases. **pollucheck** analyze different type of data-sets. I have been working with air pollution since 14 years ago in many countries and I`m sure this software will be used in many countries. Thanks for the invitation to review this paper and have a nice day.
