### 2021-07-19 adithirgis/OpenSourceAirQualityApp#1 @adithirgis
Oh thanks so much, we have addressed the OpenAQ issue, please let us know if you still face same issue.
