### 2021-07-19 #3435 @adithirgis
Thank you so much @ibarraespinosa , we have addressed the OpenAQ issue, please let us know if you still face same issue.
