### 2021-07-20 #3435 @adithirgis
Hi @ibarraespinosa!

We have implemented your suggestion of adding case study.

Thank you!
