### 2021-07-20 #3435 @ibarraespinosa
Hi Adithi, thanks you!.

Em ter., 20 de jul. de 2021 às 08:31, Adithi R. Upadhya <
***@***.***> escreveu:

> Hi @ibarraespinosa <https://github.com/ibarraespinosa>!
>
> We have implemented your suggestion of adding case study.
>
> Thank you!
>
