### 2021-07-21 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
JOSS review
Hi @adithirgis. I'm opening this issue as part of my review for JOSS (openjournals/joss-reviews#3435). You can find my initial comments below. I'll go through your submission again once you make some changes and let you know if I have any further comments. Please feel free to ping me if anything is unclear.

Thanks!

## General comments

You seem to have renamed this repository from OpenSourceAirQualityApp to pollucheck. However, in a number of files, the repository's name and URL still point to the old name. These need to be fixed. The Shiny app is still called OpenSourceAirQualityApp as well, but this is not an issue, so it's up to you if you want to rename it.

Some wording should be edited for consistency. Use either:

- "open source" or "open-source".
- "visualize/visualization" or "visualize/visualisation".
- "data set" or "dataset".
- "analyze" or "analyse".

Minor corrections:

- Remove hyphens from "air-quality" and "air-pollution".
- In `.github/workflows/R-CMD-check.yaml`, the `master` branch is redundant and can be removed.

## README

In line 9, state what CPCB stands for. In line 11, I think it will be useful to add a hyperlink to the openair package.

The section headings are quite long. I think you could use shorter headings, followed by a description below. If you try to use GitHub's table of contents to navigate through the README, it will be difficult to make out the heading titles.

Some of the images are not indented correctly. Can you indent them so that they align with their corresponding bullet point?

Could you also add the installation instructions to serve the app locally here?

### Community guidelines

I think no description is needed for point 1; just add a link to CONTRIBUTING.md. I suggest adding a disclaimer that contributors must adhere to the code of conduct, and a link to it.

### Author credit statement

This is a very long sentence; do you think you can split it up?

## Contributing

The link to the Code of Conduct in CONTRIBUTING.md is broken. You can move the Code of Conduct to `.github` to fix this.

### Fixing typos

It would be good to include PR instructions in this section. You have suggested using the GitHub interface directly, so you can add a link to the relevant GitHub documentation for this (e.g., ["Creating a pull request from a fork"](https://docs.github.com/en/articles/creating-a-pull-request-from-a-fork)).

*"... as long as the changes are made in the _source_ file"*, as well as the roxygen2 description, might be a bit confusing to new contributors. If a contributor wants to fix a typo in the README, then it won't be necessary for them to learn about roxygen2 comments. My suggestion is to clarify that if contributors find any typos in an `.Rd` file, then they should make changes in the corresponding `.R` file, as `.Rd` files are automatically generated by roxygen2 and should not be edited by hand.

### Bigger changes

I think `install.packages("devtools")` should be the first step here. My base R environment did not include devtools, so I had to install it first.

You can probably remove "If you haven't done this before", since this is targeted towards those already familiar with Git and R/tidyverse. I think a link to usethis's documentation would be useful here though.

In point 2, you mention that `devtools::check()` should pass cleanly. However, I got a note regarding the package size when I ran the check:

```text
── R CMD check results ─────────────────────────────────── pollucheck 1.0.0 ────
Duration: 12.8s

❯ checking installed package size ... NOTE
    installed size is  7.6Mb
    sub-directories of 1Mb or more:
      shiny   7.5Mb

0 errors ✔ | 0 warnings ✔ | 1 note ✖
```

I don't know how this note can be fixed. Perhaps you could clarify that the check should pass without errors or warnings.

Point 4: "git" -> "Git".

Point 5: *"For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header)."* There is no `NEWS.md` at present, so you could create one and add some placeholder text.

## DESCRIPTION

Suggested changes by line number:

- [22] mention that this is specific to Indian air pollution data
- [24] github > GitHub
- [24-27] the sentence re: wind direction could use some grammar fixes
- [26] time-resolution -> time resolution

## App

These things were picked up by the linter in `inst/shiny/app.R`:

- [933, 935, 937] Use <-, not =, for assignment.
- [950] local variable ‘R_sq’ assigned but may not be used.
- [2017, 2019] fortify {ggplot2}: "Rather than using this function, I now recommend using the broom package, which implements a much wider range of methods. `fortify()` may be deprecated in the future."

## Paper

### Summary

In the second sentence of paragraph 1, there is a comma missing between "actions" and "quantification".

In the second paragraph, I think you need to introduce pollucheck first instead of shiny. Something along the lines of "pollucheck is an interactive R application built using the shiny package...", where you add a citation to the shiny package.

I think it's important to mention in the summary that this app is for exploring **Indian** air pollution data.

You mention in the "App Display" section the different packages used for building pollucheck. I think you can move that sentence to the summary, so that this section focusses on the app's functionality.

### Limitations

In point 5, can you briefly elaborate why caution must be exercised regarding wind direction data? You stated in `DESCRIPTION` that only hourly data is processed correctly. Can you state that here?

### Installation

The heading level for the "Installation" seems incorrect.

I think it's useful to state here that the app is already hosted online on shinyapps.io, but users can choose to serve it locally using the installation instructions provided.

I recommend installing devtools as the first step. Additionally, if you are importing the pollucheck library, it won't be necessary to use the `pollucheck::` prefix to run the app. Therefore, I think your installation commands could be changed to one of the following:

```r
install.packages("devtools")
devtools::install_github("adithirgis/pollucheck")
library(pollucheck)
pollucheck_run()

# or...

install.packages("devtools")
devtools::install_github("adithirgis/pollucheck")
pollucheck::pollucheck_run()
```

In the last paragraph:

- remove the first "and" in the first sentence.
- change "act" to "acts" in the second sentence.

### Case Study

It will be good to include some information about the data used for the case study. You say it's based on 18 months of pollution data -- what's the study area and what are the data sources?

In paragraph 3, "Interquartile" should be "interquartile".

### Bibliography

The bibliography entries' titles are not capitalised correctly in the compiled PDF. To preserve capitalisation, you can wrap the titles in quotes. For example:

```text
title = "{Air pollution in mega cities in China}"
```

There is also some spacing after the decimal/dot in some of the titles:

- `Upadhya:2020`: "mmaqshiny v1. 0..."
- `Spandana:2021`: "PM2. 5/PM10 ratio..."
- `Chu:2015`: "...PM10-PM2. 5 relationship"
