### 2021-07-21 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Thanks for a thorough review, we will get back to you soon with the changes.
