+++ paper.bib
@@ -1,5 +1,5 @@
 @misc{Hystad:2020,
-  title = {Air pollution health impacts: the knowns and unknowns for reliable global burden calculations},
+  title = "{Air pollution health impacts: the knowns and unknowns for reliable global burden calculations}",
   author = {Hystad, Perry and Yusuf, Salim and Brauer, Michael},
   year = {2020},
   publisher = {Oxford University Press},
@@ -7,7 +7,7 @@
 }
 
 @article{Hyndman:2008,
-    title = {Automatic time series forecasting: the forecast package for {R}},
+    title = "{Automatic time series forecasting: the forecast package for {R}}",
     author = {Rob J Hyndman and Yeasmin Khandakar},
     journal = {Journal of Statistical Software},
     volume = {26},
@@ -19,7 +19,7 @@
   }
 
 @manual{Shiny:2021,
-  title = {shiny: Web Application Framework for R},
+  title = "{shiny: Web Application Framework for R}",
   author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
   year = {2021},
   note = {R package version 1.6.0},
@@ -27,7 +27,7 @@
 }
 
 @article{Carslaw:2012,
-  title = {openair --- An R package for air quality data analysis},
+  title = "{openair --- An R package for air quality data analysis}",
   author = {David C. Carslaw and Karl Ropkins},
   journal = {Environmental Modelling & Software},
   volume = {27--28},
@@ -39,7 +39,7 @@
 }
 
 @article{Salmon:2017,
-  title = {rtimicropem: an R package supporting the analysis of RTI MicroPEM output files},
+  title = "{rtimicropem: an R package supporting the analysis of RTI MicroPEM output files}",
   author = {Salmon, Ma{\"e}lle and Vakacherla, Sreekanth and Mil{\`a}, Carles and Marshall, Julian and Tonne, Cathryn},
   journal = {Journal of Open Source Software},
   volume = {2},
@@ -50,7 +50,7 @@
 }
 
 @article{Upadhya:2020,
-  title = {mmaqshiny v1. 0: R-Shiny package to explore Air-Quality Mobile-Monitoring data},
+  title = "{mmaqshiny v1.0: R-Shiny package to explore Air-Quality Mobile-Monitoring data}",
   author = {Upadhya, Adithi R and Agrawal, Pratyush and Vakacherla, Sreekanth and Kushwaha, Meenakshi},
   journal = {Journal of Open Source Software},
   volume = {5},
@@ -60,7 +60,7 @@
   doi = {10.21105/joss.02250}
 }
 @article{Spandana:2021,
-  title = {PM2. 5/PM10 ratio characteristics over urban sites of India},
+  title = "{PM2.5/PM10 ratio characteristics over urban sites of India}",
   author = {Spandana, B and Rao, S Srinivasa and Upadhya, Adithi R and Kulkarni, Padmavati and Sreekanth, V},
   journal = {Advances in Space Research},
   volume = {67},
@@ -72,7 +72,7 @@
 }
 
 @article{Chan:2008,
-  title = {Air pollution in mega cities in China},
+  title = "{Air pollution in mega cities in China}",
   author = {Chan, Chak K and Yao, Xiaohong},
   journal = {Atmospheric environment},
   volume = {42},
@@ -84,7 +84,7 @@
 }
 
 @article{Chu:2015,
-  title = {Modeling the spatio-temporal heterogeneity in the PM10-PM2. 5 relationship},
[…]
