+++ DESCRIPTION
@@ -33,7 +33,7 @@ Depends:
     R (>= 3.6.0)
 Imports: 
     biwavelet (>= 0.20.19),
-    broom (>=0.7.6),
+    broom (>= 0.7.6),
     bslib (>= 0.2.4),
     Cairo (>= 1.5.12.2),
     data.table (>= 1.14.0),
