+++ README.md
@@ -25,11 +25,11 @@ A walk through to use this app for everyone -
 
 ![](inst/shiny/WWW/CPCB_data_down_S.jpeg)
 
--   Select the Indian state from the "State Name" dropdown.
+-   Select the Indian state from the "State Name" drop-down.
 
 ![](inst/shiny/WWW/CPCB_Station.jpeg)
 
--   Now select the city for which the data needs to be downloaded using the "City Name" dropdown menu.
+-   Now select the city for which the data needs to be downloaded using the "City Name" drop-down menu.
 
 ![](inst/shiny/WWW/CPCB_Station_city.jpeg)
 
@@ -51,56 +51,56 @@ A walk through to use this app for everyone -
 
 #### Download data - OpenAQ 
 
--   Click on [OpenAQ](https://openaq.org/#/countries/IN?_k=5ecycz) in the app to download the OpenAQ data set.
+- Click on [OpenAQ](https://openaq.org/#/countries/IN?_k=5ecycz) in the app to download the OpenAQ data set.
 
-    <img src="inst/shiny/WWW/oaq_1.jpg" width="286"/>
+<img src="inst/shiny/WWW/oaq_1.jpg" width="286"/>
 
--   Now, click on the "*Download*" option on your browser.
+- Now, click on the "*Download*" option on your browser.
 
-    <img src="inst/shiny/WWW/oaq_2.jpg" width="517"/>
+<img src="inst/shiny/WWW/oaq_2.jpg" width="517"/>
 
--   In data download, you can download the data by *Locations* or by *Datasets*.
+- In data download, you can download the data by *Locations* or by *Datasets*.
 
-    <img src="inst/shiny/WWW/oaq_3.jpg" width="484"/>
+<img src="inst/shiny/WWW/oaq_3.jpg" width="484"/>
 
--   Now select the desired, "*Country*", "*City/Region*" and "*Location*" using the dropdown.
+- Now select the desired, "*Country*", "*City/Region*" and "*Location*" using the drop-down.
 
-    <img src="inst/shiny/WWW/oaq_4.jpg" width="306"/><img src="inst/shiny/WWW/oaq_5.jpg" width="103"/><img src="inst/shiny/WWW/oaq_6.jpg" width="215"/>
+<img src="inst/shiny/WWW/oaq_4.jpg" width="306"/><img src="inst/shiny/WWW/oaq_5.jpg" width="103"/><img src="inst/shiny/WWW/oaq_6.jpg" width="215"/>
 
--   To proceed further select the "*Start Date***"** and the "*End Date***"** for the data.
+- To proceed further select the "*Start Date***"** and the "*End Date***"** for the data.
 
-    <img src="inst/shiny/WWW/oaq_7.jpg" width="525"/>
+<img src="inst/shiny/WWW/oaq_7.jpg" width="525"/>
 
--   Now select which type of sensor is available at that location. Is it a "*Low-cost Sensor*" or "*Reference Grade*" monitor?
+- Now select which type of sensor is available at that location. Is it a "*Low-cost Sensor*" or "*Reference Grade*" monitor?
 
-    ![](inst/shiny/WWW/oaq_8.jpg)
+![](inst/shiny/WWW/oaq_8.jpg)
 
--   Finally, select the parameters from the list of "*Core"* and "*Additional Parameters"* which you wish to download from that particular location and click on "*Download Selection*". The file will be saved in your local disk in .csv format.
+- Finally, select the parameters from the list of "*Core"* and "*Additional Parameters"* which you wish to download from that particular location and click on "*Download Selection*". The file will be saved in your local disk in .csv format.
 
-    <img src="inst/shiny/WWW/oaq_9.jpg" width="519"/>
[…]
