+++ .github/CONTRIBUTING.md
@@ -6,21 +6,23 @@ For more detailed info about contributing to this, and other tidyverse packages,
 
 ## Fixing typos
 
+You can refer to [this](https://docs.github.com/en/github/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request-from-a-fork) for proposing changes with pull request from a fork. 
 You can fix typos, spelling mistakes, or grammatical errors in the documentation directly using the GitHub web interface, as long as the changes are made in the _source_ file. 
-This generally means you'll need to edit [roxygen2 comments](https://roxygen2.r-lib.org/articles/roxygen2.html) in an `.R`, not a `.Rd` file. 
-You can find the `.R` file that generates the `.Rd` by reading the comment in the first line.
+
+If you find any typos in an `.Rd` file, then please make changes in the corresponding `.R` file, as `.Rd` files are automatically generated by [roxygen2](https://roxygen2.r-lib.org/articles/roxygen2.html) and should not be edited by hand.
 
 ## Bigger changes
 
+The first step here for you to make any changes is to install `devtools` using `install.packages("devtools")`.
 If you want to make a bigger change, it's a good idea to first file an issue and make sure someone from the team agrees that it’s needed. 
 If you’ve found a bug, please file an issue that illustrates the bug with a minimal 
 [reprex](https://www.tidyverse.org/help/#reprex) (this will also help you write a unit test, if needed).
 
 ### Pull request process
 
-*   Fork the package and clone onto your computer. If you haven't done this before, we recommend using `usethis::create_from_github("adithirgis/pollucheck", fork = TRUE)`.
+*   Fork the package and clone onto your computer. We recommend using `usethis::create_from_github("adithirgis/pollucheck", fork = TRUE)`. For details on `usethis` click [here](https://cran.r-project.org/web/packages/usethis/index.html).
 
-*   Install all development dependencies with `devtools::install_dev_deps()`, and then make sure the package passes R CMD check by running `devtools::check()`. 
+*   Install all development dependencies with `devtools::install_dev_deps()`, and then make sure the package passes R CMD check by running `devtools::check()`. It should pass without errors and warnings. 
     If R CMD check doesn't pass cleanly, it's a good idea to ask for help before continuing. 
 *   Create a git branch for your pull request (PR). We recommend using `usethis::pr_init("brief-description-of-change")`.
 
