+++ NEWS.md
@@ -1,3 +1,5 @@
-## version 1.0                     
+# pollucheck 1.0                     
 
-pollucheck 
+## Major changes 
+
+## Bug fixes
