### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Hello @nmstreethran,

Thanks for a thorough review, I have updated all the comments you have provided.
Though there is a small clarification regarding the data used here -

- `pollucheck` is not specific to Indian air pollution data, only one source Central Pollution Control Board is specific to India, rest of the sources OpenAQ and AirNow data is available all around the globe.

Please let us know if we need to change anything.

Cheers!
Adithi
