### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
Hi @adithirgis, thank you for the quick update! I'll go through your submission again and let you know by the end of today if there is anything else to address.

Thanks for the clarification regarding the supported data sources. Your README still says that "pollucheck helps exploring the open-source Indian air quality data", and the OpenAQ and AirNow links point to data specific to India. Could you update this?

These links seem more appropriate:

- https://www.airnow.gov/international/us-embassies-and-consulates/
- https://openaq.org/#/countries
