+++ inst/shiny/tests/shinytest/test_OpenAQ.R
@@ -10,7 +10,6 @@ app$setInputs(hourly = "click")
 app$setInputs(avg_hour = "daily")
 app$setInputs(hourly = "click")
 app$setInputs(tabs1 = "2")
-app$snapshot()
 app$setInputs(tabs1 = "5")
 app$setInputs(freq = "click")
 app$setInputs(qq = "click")
