+++ README.md
@@ -4,9 +4,9 @@
 [![R-CMD-check](https://github.com/adithirgis/pollucheck/workflows/R-CMD-check/badge.svg)](https://github.com/adithirgis/pollucheck/actions)
  
 
-pollucheck helps exploring the open-source Indian air quality data.
+pollucheck helps exploring the open-source air quality data.
 
--   pollucheck allows users to handle open-source air quality data sets available from OpenAQ (<https://openaq.org/#/countries/IN?_k=5ecycz>), Central Pollution Control Board [CPCB](https://app.cpcbccr.com/ccr/#/caaqm-dashboard-all/caaqm-landing), and AirNow (<https://www.airnow.gov/international/us-embassies-and-consulates/#India>).
+-   pollucheck allows users to handle open-source air quality data sets available from OpenAQ (<https://openaq.org/#/countries>), Central Pollution Control Board [CPCB](https://app.cpcbccr.com/ccr/#/caaqm-dashboard-all/caaqm-landing), and AirNow (<https://www.airnow.gov/international/us-embassies-and-consulates/>).
 -   Users can visualize data, analyze data, perform basic statistical operations, and generate a variety of publication-ready plots.
 -   We have also included the popular [openair](https://cran.r-project.org/web/packages/openair/index.html) package in this app.
 -   We have hosted this app here - https://aruapps.shinyapps.io/OpenSourceAirQualityApp/ 
