### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Thanks so much for these as well. We have updated now the documentation as well.

Cheers!
Adithi
