### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
Thanks @adithirgis. I have just a few minor comments left:

`DESCRIPTION`

- line 26: Should be "Wind direction at any other time resolution...".

`paper.md`

- The review guidelines specifically asks for a "Statement of Need" section in the paper. Do you think you could add this? I think lines 38-47 already explain the need for the pollucheck package, so you only need to add a new heading.
- In line 51 / point i: "A set of quality check options **for the** (a)...".
- In line 55 / point iii: there should be commas between points (a) to (e). "Anderson-Darling" needs to be hyphenated.
- In lines 65 (end of sentence) and 92 (first sentence), `pollucheck` should be enclosed in backticks to be consistent with the rest of the paper.
- The "Installation" section level seems incorrect (it's currently a sub-section of "Limitations").
- Hyphens from "time series" and "quality check" should be removed.

`inst/shiny/WWW/include.Rmd`

- In lines 33 and 75, the repository URL should be updated to <https://github.com/adithirgis/pollucheck>

`inst/shiny/app.R`

- line 1155: the URL should be updated to <https://github.com/adithirgis/pollucheck#app-usage>

`README.md`

- line 70: the bold/italic formatting needs to be fixed.
- line 109: "outliers" is spelled incorrectly.
- lines 125-127: would it be more appropriate to move these to the Community Guidelines?
- line 201: this sentence seems incomplete?
