### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
@adithirgis Additionally, there were a few conflicts and a note regarding a deprecated function when I served the app locally using `pollucheck::pollucheck_run()`. You can find the log here:

```text
Loading required package: shiny
── Attaching packages ───────────────────────────────────────────────────── tidyverse 1.3.1 ──
✔ ggplot2 3.3.5     ✔ purrr   0.3.4
✔ tibble  3.1.2     ✔ dplyr   1.0.7
✔ tidyr   1.1.3     ✔ stringr 1.4.0
✔ readr   2.0.0     ✔ forcats 0.5.1
── Conflicts ──────────────────────────────────────────────────────── tidyverse_conflicts() ──
✖ dplyr::filter() masks stats::filter()
✖ dplyr::lag()    masks stats::lag()
Stackoverflow is a great place to get help:
        https://stackoverflow.com/tags/shinyjs

Attaching package: ‘shinyjs’

The following object is masked from ‘package:shiny’:

    runExample

The following objects are masked from ‘package:methods’:

    removeClass, show

Registered S3 method overwritten by 'quantmod':
  method            from
  as.zoo.data.frame zoo
biwavelet 0.20.21 loaded.

Attaching package: ‘biwavelet’

The following object is masked from ‘package:ggplot2’:

    arrow

Attaching package: ‘DT’

The following objects are masked from ‘package:shiny’:

    dataTableOutput, renderDataTable

Attaching package: ‘broom’

The following object is masked from ‘package:bslib’:

    bootstrap

data.table 1.14.0 using 2 threads (see ?getDTthreads).  Latest news: r-datatable.com

Attaching package: ‘data.table’

The following objects are masked from ‘package:dplyr’:

    between, first, last

The following object is masked from ‘package:purrr’:

    transpose

Attaching package: ‘janitor’

The following objects are masked from ‘package:stats’:

    chisq.test, fisher.test

Attaching package: ‘zoo’

The following objects are masked from ‘package:base’:

    as.Date, as.Date.numeric

Listening on http://127.0.0.1:3542
New names:
* `` -> ...3
* `` -> ...4
* `` -> ...5
* `` -> ...6
Warning in FUN(X[[i]], ...) : NAs introduced by coercion
Warning in FUN(X[[i]], ...) : NAs introduced by coercion
Warning in FUN(X[[i]], ...) : NAs introduced by coercion
Warning in FUN(X[[i]], ...) : NAs introduced by coercion
`mutate_all()` ignored the following grouping variables:
Columns `day`, `Location`
Use `mutate_at(df, vars(-group_cols()), myoperation)` to silence the message.
Warning: `funs()` was deprecated in dplyr 0.8.0.
Please use a list of either functions or lambdas:

  # Simple named list:
  list(mean = mean, median = median)

  # Auto named with `tibble::lst()`:
  tibble::lst(mean, median)

  # Using lambdas
  list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))
This warning is displayed once every 8 hours.
Call `lifecycle::last_warnings()` to see where this warning was generated.
```
