### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
This is the output of `lifecycle::last_warnings()`:

```text
[[1]]
<deprecated>
message: `funs()` was deprecated in dplyr 0.8.0.
Please use a list of either functions or lambdas:

  # Simple named list:
  list(mean = mean, median = median)

  # Auto named with `tibble::lst()`:
  tibble::lst(mean, median)

  # Using lambdas
  list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))
Backtrace:
  1. pollucheck::pollucheck_run()
 72. dplyr::funs(mean, sd)
 ```
