+++ inst/shiny/app.R
@@ -764,7 +764,7 @@ server <- function(input, output, session) {
     all$hour <- lubridate::ceiling_date(all$date, "hour")
     all <- all %>%
       group_by(hour) %>%
-      summarise_all(funs(mean), na.rm = TRUE) %>%
+      summarise_all(list(~ mean(., na.rm = TRUE))) %>%
       dplyr::select(everything(),-date) %>%
       dplyr::select("date" = hour, everything())
     names(all) <- toupper(names(all))
@@ -865,7 +865,7 @@ server <- function(input, output, session) {
   outlier <- function(site1_join_f1, name, date, eq) {
     site1_join_f1 <- site1_join_f1 %>%
       group_by(day, Location) %>%
-      dplyr::mutate_all(funs(mean, sd), na.rm = TRUE) %>%
+      dplyr::mutate_all(list(~ mean(., na.rm = TRUE), ~ sd(., na.rm = TRUE))) %>%
       ungroup() %>%
       dplyr::select(date, day, Location, everything(),-date_mean,-date_sd)
     tseries_df <- site1_join_f1 %>%
@@ -1152,7 +1152,7 @@ server <- function(input, output, session) {
     ),
     footer = tagList(
       actionButton("help_link", "Read Me",
-                   onclick = "window.open('https://github.com/adithirgis/pollucheck#have-a-look-at-the-app',
+                   onclick = "window.open('https://github.com/adithirgis/pollucheck#app-usage',
                    '_blank')"),
       actionButton("switch_tab", "FAQ's"),
       modalButton("Close")
@@ -1460,7 +1460,7 @@ server <- function(input, output, session) {
           )
         ) %>%
         group_by(date, Location) %>%
-        summarise_all(funs(mean), na.rm = TRUE)
+        summarise_all(list(~ mean(., na.rm = TRUE)))
       x <- zoo(data[[input$palleInp2]], data$date)
       x <- na.interp(x)
     } else {
@@ -1507,7 +1507,7 @@ server <- function(input, output, session) {
       data <- data %>%
         dplyr::select(hour, y) %>%
         group_by(hour) %>%
-        summarise_all(funs(mean, sd), na.rm = TRUE)
+        summarise_all(list(~ mean(., na.rm = TRUE), ~ sd(., na.rm = TRUE)))
       ggplot(data, aes(hour, mean)) + geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
                                                     color = "seagreen") +
         scale_x_continuous(limits = c(-1, 24),
@@ -1522,7 +1522,8 @@ server <- function(input, output, session) {
       data <- data %>%
         dplyr::select(hour, y) %>%
         group_by(hour) %>%
-        summarise_all(funs(median, p25 = quantile(., .25), p75 = quantile(., .75)), na.rm = TRUE)
+        summarise_all(list(median = ~ median(.x, na.rm = TRUE), p25 = ~ quantile(.x, .25, na.rm = TRUE), 
+                           p75 = ~ quantile(.x, .75, na.rm = TRUE)))
       ggplot(data, aes(hour, median)) + geom_errorbar(aes(ymin = p25, ymax = p75),
                                                       color = "seagreen") +
         scale_x_continuous(limits = c(-1, 24),
@@ -1536,7 +1537,8 @@ server <- function(input, output, session) {
     } else if (input$diur == "mediq" && input$diurn == "mon") {
       data <- data %>%
         group_by(month, hour) %>%
-        summarise_all(funs(median, p25 = quantile(., .25), p75 = quantile(., .75)), na.rm = TRUE)
[…]
