### 2021-07-22 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Hello @nmstreethran!

Thank so much for pointing the deprecated function.

> README.md
line 109: "outliers" is spelled incorrectly.

I am not able to get this wrong spelling could you please point it out? Is it spelled outliers?

Thanks again
Cheers!
Adithi
