### 2021-07-23 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Hello @nmstreethran!

We have implemented all your suggestions.

Thanks!
Adithi
