### 2021-07-23 adithirgis/OpenSourceAirQualityApp#2 @nmstreethran
Hi @adithirgis, thank you very much!

Here is the spelling mistake in README (line 110):

> Specify a multiple (X) to remove outliers based on Mean and SD- If you want to clean your data set based on outliers, not usually necessary, use only if you want to remove **outiliers** based on Mean and Standard Deviation values.

You may close this issue once you correct this typo. I'll inform the editor that my review is complete.

Great work on the app!
