### 2021-07-23 #3435 @nmstreethran
Hi @ajstewartlang, I have completed my review (adithirgis/pollucheck#2). `pollucheck` is very useful for performing exploratory data analysis on open air pollution data. The app provides users an interactive way to do this without needing any programming knowledge. I'm happy to recommend this submission for publication. Please let me know if there is anything else I can do.

Thanks!
