### 2021-07-23 adithirgis/OpenSourceAirQualityApp#2 @adithirgis
Hello @nmstreethran,

Thanks for reviewing and helping us improve this application. I have fixed the typo.

Cheers!
Adithi

Do we need to close this issue?
