### 2021-07-23 #3435 @ajstewartlang
Many thanks @ibarraespinosa and @nmstreethran for your helpful and detailed reviews. I've had a read through myself @adithirgis and agree that `pollucheck` is a very nice and useful piece of software. Congrats!

@adithirgis can I now ask you to do the following please?

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
