### 2021-07-23 #3435 @adithirgis
Thanks @ajstewartlang!

 - We made a tagged release of the v1.0.0
 - Here is the link to the metadata - https://zenodo.org/record/5126939#.YPqqUegzZPY (I checked the deposit)
 - DOI version from Zenodo - 10.5281/zenodo.5126939.

 Let me know if there is any issue with these.

 Thanks again!
 Adithi
