### 2021-07-23 #3435 @danielskatz
👋 @adithirgis  - I've suggested a few small changes to the paper in https://github.com/adithirgis/pollucheck/pull/3 - please merge this, or let me know what you disagree with, then we can proceed to final acceptance and publishing.
