+++ DESCRIPTION
@@ -18,11 +18,12 @@ Authors@R:
              family = "Kushwaha",
              role = c("aut", "cph"),
              email = "meenakshi@ilklabs.com"))
-Description: pollucheck helps to analyze and visualize open-source air
-    pollution data available. An example data set is already loaded to help
-    you walk through all the features of the app. If you need help with
-    downloading your own, check the README file on GitHub. This app can
-    process all parameters except wind direction which when downloaded at
+Description: Air pollution leads to various health hazards, hence there is a need to process 
+    and effectively work with the available open-source data. pollucheck helps 
+    to analyze and visualize open-source air pollution data available. An example 
+    data set is already loaded to help you walk through all the features of the 
+    app. If you need help with downloading your own, check the README file on GitHub. 
+    This app can process all parameters except wind direction which when downloaded at
     1 hour is processed correctly. Wind direction at any other time resolution 
     (e.g. - daily average) is not averaged correctly as it is a vector and 
     cannot be averaged using simple arithmetic mean.
