### 2021-07-23 #3435 @adithirgis
Hi @danielskatz,

Is it possible to change the Zenodo now?
