### 2021-07-23 #3435 @danielskatz
you don't need to, as we don't need the zenodo archive to have the paper at all, let alone the final version, but if you want to, that's fine - you could update it.  Let me know if you want to to, and the new DOI (and new version tag) if so
