### 2021-07-23 #3435 @danielskatz
I'm not sure what you mean by "highlighted".  When I did the accept, it captured what was in the paper.md and paper.bib at that time.  Did you make changes after that?
